conftest.py: enable command-line args passed to the TestCase instance

+++ tests/test_lib.py
@@ -3,6 +3,8 @@ import unittest
 import glob
 import shutil
 
+import pytest
+
 # Note: for performance reasons some imports are moved inside the tests
 
 from ipydex import IPS, Container
@@ -12,32 +14,39 @@ from ipydex import IPS, Container
 pjoin = os.path.join
 
 dir_of_this_file = os.path.dirname(__file__)
-TESTDATA = pjoin(dir_of_this_file, "testdata")
+TEST_DATA_DIR = pjoin(dir_of_this_file, "testdata")
 
 TEST_PREFIX = "tmp_pp_"
 
-_RAW_PNG_DIR = pjoin(TESTDATA, "raw_png")
+_RAW_PNG_DIR = pjoin(TEST_DATA_DIR, "raw_png")
 
 
 class TestCases1(unittest.TestCase):
 
-    def setUp(self) -> None:
-        self.jpg0_dir_path = pjoin(TESTDATA, "jpg0")
-        return super().setUp()
+    # this is to access custom command line options passed to pytest
+    # see also conftest.py
+    @pytest.fixture(autouse=True)
+    def setup(self, keep_data, no_parallel):
+        # this is not to be confused with TestCase.setUp
+        self.keep_data = keep_data
+        self.no_parallel = no_parallel
 
     def tearDown(self):
+        if self.keep_data:
+            return
+
         dir_list = self.get_tmp_data_dirs()
         for dir in dir_list:
             shutil.rmtree(dir, ignore_errors=True)
 
     def get_tmp_data_dirs(self):
-        dir_list = glob.glob(f"{TEST_PREFIX}*")
+        dir_list = glob.glob(pjoin(TEST_DATA_DIR, f"{TEST_PREFIX}*"))
         return dir_list
 
     def get_png_dir_path(self):
         png_pattern = pjoin(_RAW_PNG_DIR, "*.png")
         png_files = glob.glob(png_pattern)
-        jpg_pattern = pjoin(TESTDATA, "_jpg_templates", "*.jpg")
+        jpg_pattern = pjoin(TEST_DATA_DIR, "_jpg_templates", "*.jpg")
         jpg_files = glob.glob(jpg_pattern)
         if len(png_files) < len(jpg_files):
             # create png files from jpg files (needs to be run only once)
@@ -84,19 +93,19 @@ class TestCases1(unittest.TestCase):
     def test020__bboxes(self):
         from chimcla import stage_2a_bar_selection as bs
 
-        tmp_path = pjoin(TESTDATA, "stage1_completed", "2023-06-26_06-19-58_C50.jpg")
+        tmp_path = pjoin(TEST_DATA_DIR, "stage1_completed", "2023-06-26_06-19-58_C50.jpg")
         ccia = bs.CavityCarrierImageAnalyzer(tmp_path, bboxes=True)
 
     def test030__symloghist(self):
         from chimcla import stage_2a_bar_selection as bs
 
-        tmp_path = pjoin(TESTDATA, "stage1_completed", "2023-06-26_06-19-58_C50.jpg")
+        tmp_path = pjoin(TEST_DATA_DIR, "stage1_completed", "2023-06-26_06-19-58_C50.jpg")
         bs.get_symlog_hist(tmp_path, *"a 20".split(), dc=None)
 
     def test040__find_critical(self):
         from chimcla import stage_2a_bar_selection as bs
 
-        tmp_path = pjoin(TESTDATA, "stage1_completed", "2023-06-26_06-19-25_C50.jpg")
+        tmp_path = pjoin(TEST_DATA_DIR, "stage1_completed", "2023-06-26_06-19-25_C50.jpg")
         he = bs.HistEvaluation(img_fpath=tmp_path)
         he.initialize_hist_cache()
         he.find_critical_cells_for_img(save_options={"save_plot": False, "push_db": False})
